requirements.txt
requests>=2.32.0
pandas>=2.2.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
kafka-python>=2.0.0
streamlit>=1.32.0
python-dotenv>=1.0.0
pytest>=8.0.0
